src/app/dashboard/features/dashboard.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';

import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
import { DashboardComponent } from './dashboard.component';
import { By } from '@angular/platform-browser';
import { HttpClientTestingModule } from '@angular/common/http/testing';

describe('DashboardComponent', () => {
  let component: DashboardComponent;
  let fixture: ComponentFixture<DashboardComponent>;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      imports: [
        BrowserAnimationsModule,
        DashboardComponent,
        HttpClientTestingModule,
      ],
    }).compileComponents();

    fixture = TestBed.createComponent(DashboardComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  describe('DOM', () => {
    it('should have a sidebar element and a content element', () => {
      const sidenav = fixture.debugElement.query(
        By.css('[data-testid=sidenav]'),
      );

      expect(sidenav).toBeTruthy();

      const content = fixture.debugElement.query(
        By.css('[data-testid=content]'),
      );

      expect(content).toBeTruthy();
    });
  });
});
